+++ src/main/java/gregtech/api/items/armor/ArmorMetaItem.java
@@ -1,6 +1,6 @@
 package gregtech.api.items.armor;
 
-import gregtech.api.creativetab.GTCreativeTabs;
+import gregtech.common.creativetab.GTCreativeTabs;
 import gregtech.api.items.metaitem.MetaItem;
 import gregtech.api.items.metaitem.stats.IEnchantabilityHelper;
 import gregtech.api.items.metaitem.stats.IItemComponent;
