+++ src/main/java/gregtech/common/blocks/BlockCompressed.java
@@ -1,6 +1,6 @@
 package gregtech.common.blocks;
 
-import gregtech.api.creativetab.GTCreativeTabs;
+import gregtech.common.creativetab.GTCreativeTabs;
 import gregtech.api.items.toolitem.ToolClasses;
 import gregtech.api.unification.material.Material;
 import gregtech.api.unification.material.info.MaterialIconType;
