+++ src/main/java/gregtech/common/blocks/BlockLamp.java
@@ -1,7 +1,7 @@
 package gregtech.common.blocks;
 
 import gregtech.api.GTValues;
-import gregtech.api.creativetab.GTCreativeTabs;
+import gregtech.common.creativetab.GTCreativeTabs;
 import gregtech.api.unification.OreDictUnifier;
 import gregtech.api.unification.material.MarkerMaterials;
 import gregtech.api.unification.ore.OrePrefix;
