+++ src/main/java/gregtech/common/blocks/BlockWarningSign.java
@@ -1,7 +1,7 @@
 package gregtech.common.blocks;
 
 import gregtech.api.block.VariantBlock;
-import gregtech.api.creativetab.GTCreativeTabs;
+import gregtech.common.creativetab.GTCreativeTabs;
 import gregtech.api.items.toolitem.ToolClasses;
 
 import net.minecraft.block.SoundType;
