+++ src/main/java/gregtech/common/blocks/wood/BlockRubberLeaves.java
@@ -1,6 +1,6 @@
 package gregtech.common.blocks.wood;
 
-import gregtech.api.creativetab.GTCreativeTabs;
+import gregtech.common.creativetab.GTCreativeTabs;
 import gregtech.common.blocks.MetaBlocks;
 import gregtech.core.CoreModule;
