+++ src/main/java/gregtech/common/blocks/wood/BlockRubberSapling.java
@@ -1,6 +1,6 @@
 package gregtech.common.blocks.wood;
 
-import gregtech.api.creativetab.GTCreativeTabs;
+import gregtech.common.creativetab.GTCreativeTabs;
 import gregtech.common.worldgen.WorldGenRubberTree;
 
 import net.minecraft.block.BlockBush;
