+++ src/main/java/gregtech/common/pipelike/laser/BlockLaserPipe.java
@@ -1,7 +1,7 @@
 package gregtech.common.pipelike.laser;
 
 import gregtech.api.capability.GregtechTileCapabilities;
-import gregtech.api.creativetab.GTCreativeTabs;
+import gregtech.common.creativetab.GTCreativeTabs;
 import gregtech.api.items.toolitem.ToolClasses;
 import gregtech.api.items.toolitem.ToolHelper;
 import gregtech.api.pipenet.block.BlockPipe;
